+++ src/background/background.js
@@ -1,25 +1,24 @@
 // log some info
-browser.runtime.getPlatformInfo().then(async platformInfo => {
-  const browserInfo = browser.runtime.getBrowserInfo ? await browser.runtime.getBrowserInfo() : "Can't get browser info"
+browser.runtime.getPlatformInfo().then(async (platformInfo) => {
+  const browserInfo = browser.runtime.getBrowserInfo ? await browser.runtime.getBrowserInfo() : "Can't get browser info";
   console.info(platformInfo, browserInfo);
 });
 
 // add notification listener for foreground page messages
 browser.runtime.onMessage.addListener(notify);
 // create context menus
-createMenus()
+createMenus();
 
 TurndownService.prototype.defaultEscape = TurndownService.prototype.escape;
 
 // function to convert the article content to markdown using Turndown
 function turndown(content, options, article) {
-
   if (options.turndownEscape) TurndownService.prototype.escape = TurndownService.prototype.defaultEscape;
-  else TurndownService.prototype.escape = s => s;
+  else TurndownService.prototype.escape = (s) => s;
 
   var turndownService = new TurndownService(options);
 
-  turndownService.use(turndownPluginGfm.gfm)
+  turndownService.use(turndownPluginGfm.gfm);
 
   turndownService.keep(['iframe', 'sub', 'sup', 'u', 'ins', 'del', 'small', 'big']);
 
@@ -29,12 +28,11 @@ function turndown(content, options, article) {
     filter: function (node, tdopts) {
       // if we're looking at an img node with a src
       if (node.nodeName == 'IMG' && node.getAttribute('src')) {
-        
         // get the original src
-        let src = node.getAttribute('src')
+        let src = node.getAttribute('src');
         // set the new src
         node.setAttribute('src', validateUri(src, article.baseURI));
-        
+
         // if we're downloading images, there's more to do.
         if (options.downloadImages) {
           // generate a file name for the image
@@ -52,20 +50,23 @@ function turndown(content, options, article) {
             imageList[src] = imageFilename;
           }
           // check if we're doing an obsidian style link
-          const obsidianLink = options.imageStyle.startsWith("obsidian");
+          const obsidianLink = options.imageStyle.startsWith('obsidian');
           // figure out the (local) src of the image
-          const localSrc = options.imageStyle === 'obsidian-nofolder'
-            // if using "nofolder" then we just need the filename, no folder
-            ? imageFilename.substring(imageFilename.lastIndexOf('/') + 1)
-            // otherwise we may need to modify the filename to uri encode parts for a pure markdown link
-            : imageFilename.split('/').map(s => obsidianLink ? s : encodeURI(s)).join('/')
-          
+          const localSrc =
+            options.imageStyle === 'obsidian-nofolder'
+              ? // if using "nofolder" then we just need the filename, no folder
+                imageFilename.substring(imageFilename.lastIndexOf('/') + 1)
+              : // otherwise we may need to modify the filename to uri encode parts for a pure markdown link
+                imageFilename
+                  .split('/')
+                  .map((s) => (obsidianLink ? s : encodeURI(s)))
+                  .join('/');
+
           // set the new src attribute to be the local filename
-          if(options.imageStyle != 'originalSource' && options.imageStyle != 'base64') node.setAttribute('src', localSrc);
+          if (options.imageStyle != 'originalSource' && options.imageStyle != 'base64') node.setAttribute('src', localSrc);
           // pass the filter if we're making an obsidian link (or stripping links)
           return true;
-        }
-        else return true
+        } else return true;
       }
       // don't pass the filter, just output a normal markdown link
       return false;
@@ -85,8 +86,7 @@ function turndown(content, options, article) {
           var id = this.references.length + 1;
           this.references.push('[fig' + id + ']: ' + src + titlePart);
           return '![' + alt + '][fig' + id + ']';
-        }
-        else return src ? '![' + alt + ']' + '(' + src + titlePart + ')' : ''
+        } else return src ? '![' + alt + ']' + '(' + src + titlePart + ')' : '';
       }
     },
     references: [],
@@ -96,9 +96,8 @@ function turndown(content, options, article) {
         references = '\n\n' + this.references.join('\n') + '\n\n';
         this.references = []; // Reset references
       }
-      return references
-    }
-
+      return references;
+    },
   });
 
   // add a rule for links
@@ -117,7 +116,7 @@ function turndown(content, options, article) {
       return false;
     },
     // if the filter passes, we're stripping links, so just return the content
-    replacement: (content, node, tdopts) => content
+    replacement: (content, node, tdopts) => content,
   });
 
   // handle multiple lines math
@@ -132,10 +131,8 @@ function turndown(content, options, article) {
       if (math.inline) {
         tex = tex.replaceAll('\n', ' ');
         return `$${tex}$`;
-      }
-      else
-        return `$$\n${tex}\n$$`;
-    }
+      } else return `$$\n${tex}\n$$`;
+    },
   });
 
   function repeat(character, count) {
@@ -162,69 +159,56 @@ function turndown(content, options, article) {
 
     const fence = repeat(fenceChar, fenceSize);
 
-    return (
-      '\n\n' + fence + language + '\n' +
-      code.replace(/\n$/, '') +
-      '\n' + fence + '\n\n'
-    )
+    return '\n\n' + fence + language + '\n' + code.replace(/\n$/, '') + '\n' + fence + '\n\n';
   }
 
   turndownService.addRule('fencedCodeBlock', {
     filter: function (node, options) {
-      return (
-        options.codeBlockStyle === 'fenced' &&
-        node.nodeName === 'PRE' &&
-        node.firstChild &&
-        node.firstChild.nodeName === 'CODE'
-      );
+      return options.codeBlockStyle === 'fenced' && node.nodeName === 'PRE' && node.firstChild && node.firstChild.nodeName === 'CODE';
     },
     replacement: function (content, node, options) {
       return convertToFencedCodeBlock(node.firstChild, options);
-    }
+    },
   });
 
   // handle <pre> as code blocks
   turndownService.addRule('pre', {
     filter: (node, tdopts) => {
-      return node.nodeName == 'PRE'
-             && (!node.firstChild || node.firstChild.nodeName != 'CODE')
-             && !node.querySelector('img');
+      return node.nodeName == 'PRE' && (!node.firstChild || node.firstChild.nodeName != 'CODE') && !node.querySelector('img');
     },
     replacement: (content, node, tdopts) => {
       return convertToFencedCodeBlock(node, tdopts);
-    }
+    },
   });
 
-  let markdown = options.frontmatter + turndownService.turndown(content)
-      + options.backmatter;
+  let markdown = options.frontmatter + turndownService.turndown(content) + options.backmatter;
 
   // strip out non-printing special characters which CodeMirror displays as a red dot
   // see: https://codemirror.net/doc/manual.html#option_specialChars
   markdown = markdown.replace(/[\u0000-\u0008\u000b\u000c\u000e-\u001f\u007f-\u009f\u00ad\u061c\u200b-\u200f\u2028\u2029\ufeff\ufff9-\ufffc]/g, '');
-  
+
   return { markdown: markdown, imageList: imageList };
 }
 
 function cleanAttribute(attribute) {
-  return attribute ? attribute.replace(/(\n+\s*)+/g, '\n') : ''
+  return attribute ? attribute.replace(/(\n+\s*)+/g, '\n') : '';
 }
 
 function validateUri(href, baseURI) {
   // check if the href is a valid url
   try {
     new URL(href);
-  }
-  catch {
+  } catch {
     // if it's not a valid url, that likely means we have to prepend the base uri
     const baseUri = new URL(baseURI);
 
     // if the href starts with '/', we need to go from the origin
     if (href.startsWith('/')) {
-      href = baseUri.origin + href
+      href = baseUri.origin + href;
     }
     // otherwise we need to go from the local folder
     else {
-      href = baseUri.href + (baseUri.href.endsWith('/') ? '/' : '') + href
+      href = baseUri.href + (baseUri.href.endsWith('/') ? '/' : '') + href;
     }
   }
   return href;
@@ -235,20 +219,19 @@ function getImageFilename(src, options, prependFilePath = true) {
   const queryPos = src.indexOf('?');
   let filename = src.substring(slashPos + 1, queryPos > 0 ? queryPos : src.length);
 
-  let imagePrefix = (options.imagePrefix || '');
+  let imagePrefix = options.imagePrefix || '';
 
   if (prependFilePath && options.title.includes('/')) {
     imagePrefix = options.title.substring(0, options.title.lastIndexOf('/') + 1) + imagePrefix;
+  } else if (prependFilePath) {
+    imagePrefix = options.title + (imagePrefix.startsWith('/') ? '' : '/') + imagePrefix;
   }
-  else if (prependFilePath) {
-    imagePrefix = options.title + (imagePrefix.startsWith('/') ? '' : '/') + imagePrefix
-  }
-  
+
   if (filename.includes(';base64,')) {
     // this is a base64 encoded image, so what are we going to do for a filename here?
     filename = 'image.' + filename.substring(0, filename.indexOf(';'));
   }
-  
+
   let extension = filename.substring(filename.lastIndexOf('.'));
   if (extension == filename) {
     // there is no extension, so we need to figure one out
@@ -264,11 +247,12 @@ function getImageFilename(src, options, prependFilePath = true) {
 // function to replace placeholder strings with article info
 function textReplace(string, article, disallowedChars = null) {
   for (const key in article) {
-    if (article.hasOwnProperty(key) && key != "content") {
+    if (article.hasOwnProperty(key) && key != 'content') {
       let s = (article[key] || '') + '';
       if (s && disallowedChars) s = this.generateValidFileName(s, disallowedChars);
 
-      string = string.replace(new RegExp('{' + key + '}', 'g'), s)
+      string = string
+        .replace(new RegExp('{' + key + '}', 'g'), s)
         .replace(new RegExp('{' + key + ':lower}', 'g'), s.toLowerCase())
         .replace(new RegExp('{' + key + ':upper}', 'g'), s.toUpperCase())
         .replace(new RegExp('{' + key + ':kebab}', 'g'), s.replace(/ /g, '-').toLowerCase())
@@ -276,18 +260,24 @@ function textReplace(string, article, disallowedChars = null) {
         .replace(new RegExp('{' + key + ':snake}', 'g'), s.replace(/ /g, '_').toLowerCase())
         .replace(new RegExp('{' + key + ':mixed_snake}', 'g'), s.replace(/ /g, '_'))
         // For Obsidian Custom Attachment Location plugin, we need to replace spaces with hyphens, but also remove any double hyphens.
-        .replace(new RegExp('{' + key + ':obsidian-cal}', 'g'), s.replace(/ /g, '-').replace(/-{2,}/g, "-"))
-        .replace(new RegExp('{' + key + ':camel}', 'g'), s.replace(/ ./g, (str) => str.trim().toUpperCase()).replace(/^./, (str) => str.toLowerCase()))
-        .replace(new RegExp('{' + key + ':pascal}', 'g'), s.replace(/ ./g, (str) => str.trim().toUpperCase()).replace(/^./, (str) => str.toUpperCase()))
+        .replace(new RegExp('{' + key + ':obsidian-cal}', 'g'), s.replace(/ /g, '-').replace(/-{2,}/g, '-'))
+        .replace(
+          new RegExp('{' + key + ':camel}', 'g'),
+          s.replace(/ ./g, (str) => str.trim().toUpperCase()).replace(/^./, (str) => str.toLowerCase())
+        )
+        .replace(
+          new RegExp('{' + key + ':pascal}', 'g'),
+          s.replace(/ ./g, (str) => str.trim().toUpperCase()).replace(/^./, (str) => str.toUpperCase())
+        );
     }
   }
 
   // replace date formats
   const now = new Date();
-  const dateRegex = /{date:(.+?)}/g
+  const dateRegex = /{date:(.+?)}/g;
   const matches = string.match(dateRegex);
   if (matches && matches.forEach) {
-    matches.forEach(match => {
+    matches.forEach((match) => {
       const format = match.substring(6, match.length - 1);
       const dateString = moment(now).format(format);
       string = string.replaceAll(match, dateString);
@@ -295,23 +285,22 @@ function textReplace(string, article, disallowedChars = null) {
   }
 
   // replace keywords
-  const keywordRegex = /{keywords:?(.*)?}/g
+  const keywordRegex = /{keywords:?(.*)?}/g;
   const keywordMatches = string.match(keywordRegex);
   if (keywordMatches && keywordMatches.forEach) {
-    keywordMatches.forEach(match => {
-      let seperator = match.substring(10, match.length - 1)
+    keywordMatches.forEach((match) => {
+      let seperator = match.substring(10, match.length - 1);
       try {
         seperator = JSON.parse(JSON.stringify(seperator).replace(/\\\\/g, '\\'));
-      }
-      catch { }
+      } catch {}
       const keywordsString = (article.keywords || []).join(seperator);
       string = string.replace(new RegExp(match.replace(/\\/g, '\\\\'), 'g'), keywordsString);
-    })
+    });
   }
 
   // replace anything left in curly braces
-  const defaultRegex = /{(.*?)}/g
-  string = string.replace(defaultRegex, '')
+  const defaultRegex = /{(.*?)}/g;
+  string = string.replace(defaultRegex, '');
 
   return string;
 }
@@ -327,13 +316,14 @@ async function convertArticleToMarkdown(article, downloadImages = null) {
   if (options.includeTemplate) {
     options.frontmatter = textReplace(options.frontmatter, article) + '\n';
     options.backmatter = '\n' + textReplace(options.backmatter, article);
-  }
-  else {
+  } else {
     options.frontmatter = options.backmatter = '';
   }
 
   options.imagePrefix = textReplace(options.imagePrefix, article, options.disallowedChars)
-    .split('/').map(s=>generateValidFileName(s, options.disallowedChars)).join('/');
+    .split('/')
+    .map((s) => generateValidFileName(s, options.disallowedChars))
+    .join('/');
 
   let result = turndown(article.content, options, article);
   if (options.downloadImages && options.downloadMode == 'downloadsApi') {
@@ -347,14 +337,16 @@ async function convertArticleToMarkdown(article, downloadImages = null) {
 function generateValidFileName(title, disallowedChars = null) {
   if (!title) return title;
   else title = title + '';
-  // remove < > : " / \ | ? * 
+  // remove < > : " / \ | ? *
   var illegalRe = /[\/\?<>\\:\*\|":]/g;
   // and non-breaking spaces (thanks @Licat)
-  var name = title.replace(illegalRe, "").replace(new RegExp('\u00A0', 'g'), ' ')
-      // collapse extra whitespace
-      .replace(new RegExp(/\s+/, 'g'), ' ')
-      // remove leading/trailing whitespace that can cause issues when using {pageTitle} in a download path
-      .trim();
+  var name = title
+    .replace(illegalRe, '')
+    .replace(new RegExp('\u00A0', 'g'), ' ')
+    // collapse extra whitespace
+    .replace(new RegExp(/\s+/, 'g'), ' ')
+    // remove leading/trailing whitespace that can cause issues when using {pageTitle} in a download path
+    .trim();
 
   if (disallowedChars) {
     for (let c of disallowedChars) {
@@ -362,68 +354,39 @@ function generateValidFileName(title, disallowedChars = null) {
       name = name.replace(new RegExp(c, 'g'), '');
     }
   }
-  
+
   return name;
 }
 
 async function preDownloadImages(imageList, markdown) {
   const options = await getOptions();
   let newImageList = {};
-  // originally, I was downloading the markdown file first, then all the images
-  // however, in some cases we need to download images *first* so we can get the
-  // proper file extension to put into the markdown.
-  // so... here we are waiting for all the downloads and replacements to complete
-  await Promise.all(Object.entries(imageList).map(([src, filename]) => new Promise((resolve, reject) => {
-        // we're doing an xhr so we can get it as a blob and determine filetype
-        // before the final save
-        const xhr = new XMLHttpRequest();
-        xhr.open('GET', src);
-        xhr.responseType = "blob";
-        xhr.onload = async function () {
-          // here's the returned blob
-          const blob = xhr.response;
 
-          if (options.imageStyle == 'base64') {
-            var reader = new FileReader();
-            reader.onloadend = function () {
-              markdown = markdown.replaceAll(src, reader.result)
-              resolve()
+  await Promise.all(
+    Object.entries(imageList).map(
+      ([src, filename]) =>
+        new Promise(async (resolve, reject) => {
+          try {
+            const result = await ImageHandler.downloadImage(src, filename);
+            if (options.imageStyle == 'base64') {
+              const reader = new FileReader();
+              reader.onloadend = function () {
+                markdown = markdown.replaceAll(src, reader.result);
+                resolve();
+              };
+              reader.readAsDataURL(result);
+            } else {
+              const blobUrl = URL.createObjectURL(result);
+              newImageList[blobUrl] = filename;
+              resolve();
             }
-            reader.readAsDataURL(blob);
+          } catch (error) {
+            console.error('Failed to download image:', error);
+            reject(error);
           }
-          else {
-
-            let newFilename = filename;
-            if (newFilename.endsWith('.idunno')) {
-              // replace any unknown extension with a lookup based on mime type
-              newFilename = filename.replace('.idunno', '.' + mimedb[blob.type]);
-
-              // and replace any instances of this in the markdown
-              // remember to url encode for replacement if it's not an obsidian link
-              if (!options.imageStyle.startsWith("obsidian")) {
-                markdown = markdown.replaceAll(filename.split('/').map(s => encodeURI(s)).join('/'), newFilename.split('/').map(s => encodeURI(s)).join('/'))
-              }
-              else {
-                markdown = markdown.replaceAll(filename, newFilename)
-              }
-            }
-
-            // create an object url for the blob (no point fetching it twice)
-            const blobUrl = URL.createObjectURL(blob);
-
-            // add this blob into the new image list
-            newImageList[blobUrl] = newFilename;
-
-            // resolve this promise now
-            // (the file might not be saved yet, but the blob is and replacements are complete)
-            resolve();
-          }
-        };
-        xhr.onerror = function () {
-          reject('A network error occurred attempting to download ' + src);
-        };
-        xhr.send();
-  })));
+        })
+    )
+  );
 
   return { imageList: newImageList, markdown: markdown };
 }
@@ -432,23 +395,23 @@ async function preDownloadImages(imageList, markdown) {
 async function downloadMarkdown(markdown, title, tabId, imageList = {}, mdClipsFolder = '') {
   // get the options
   const options = await getOptions();
-  
+
   // download via the downloads API
   if (options.downloadMode == 'downloadsApi' && browser.downloads) {
-    
     // create the object url with markdown data as a blob
-    const url = URL.createObjectURL(new Blob([markdown], {
-      type: "text/markdown;charset=utf-8"
-    }));
-  
-    try {
+    const url = URL.createObjectURL(
+      new Blob([markdown], {
+        type: 'text/markdown;charset=utf-8',
+      })
+    );
 
-      if(mdClipsFolder && !mdClipsFolder.endsWith('/')) mdClipsFolder += '/';
+    try {
+      if (mdClipsFolder && !mdClipsFolder.endsWith('/')) mdClipsFolder += '/';
       // start the download
       const id = await browser.downloads.download({
         url: url,
-        filename: mdClipsFolder + title + ".md",
-        saveAs: options.saveAs
+        filename: mdClipsFolder + title + '.md',
+        saveAs: options.saveAs,
       });
 
       // add a listener for the download completion
@@ -458,22 +421,21 @@ async function downloadMarkdown(markdown, title, tabId, imageList = {}, mdClipsF
       if (options.downloadImages) {
         // get the relative path of the markdown file (if any) for image path
         let destPath = mdClipsFolder + title.substring(0, title.lastIndexOf('/'));
-        if(destPath && !destPath.endsWith('/')) destPath += '/';
+        if (destPath && !destPath.endsWith('/')) destPath += '/';
         Object.entries(imageList).forEach(async ([src, filename]) => {
           // start the download of the image
           const imgId = await browser.downloads.download({
             url: src,
             // set a destination path (relative to md file)
             filename: destPath ? destPath + filename : filename,
-            saveAs: false
-          })
+            saveAs: false,
+          });
           // add a listener (so we can release the blob url)
           browser.downloads.onChanged.addListener(downloadListener(imgId, src));
         });
       }
-    }
-    catch (err) {
-      console.error("Download failed", err);
+    } catch (err) {
+      console.error('Download failed', err);
     }
   }
   // // download via obsidian://new uri
@@ -492,38 +454,37 @@ async function downloadMarkdown(markdown, title, tabId, imageList = {}, mdClipsF
   //     // the page, for example if the tab is a privileged page.
   //     console.error("Failed to execute script: " + error);
   //   };
-    
+
   // }
   // download via content link
   else {
     try {
       await ensureScripts(tabId);
-      const filename = mdClipsFolder + generateValidFileName(title, options.disallowedChars) + ".md";
-      const code = `downloadMarkdown("${filename}","${base64EncodeUnicode(markdown)}");`
-      await browser.tabs.executeScript(tabId, {code: code});
-    }
-    catch (error) {
+      const filename = mdClipsFolder + generateValidFileName(title, options.disallowedChars) + '.md';
+      const code = `downloadMarkdown("${filename}","${base64EncodeUnicode(markdown)}");`;
+      await browser.tabs.executeScript(tabId, { code: code });
+    } catch (error) {
       // This could happen if the extension is not allowed to run code in
       // the page, for example if the tab is a privileged page.
-      console.error("Failed to execute script: " + error);
-    };
+      console.error('Failed to execute script: ' + error);
+    }
   }
 }
 
 function downloadListener(id, url) {
   const self = (delta) => {
-    if (delta.id === id && delta.state && delta.state.current == "complete") {
+    if (delta.id === id && delta.state && delta.state.current == 'complete') {
       // detatch this listener
       browser.downloads.onChanged.removeListener(self);
       //release the url for the blob
       URL.revokeObjectURL(url);
     }
-  }
+  };
   return self;
 }
 
 function base64EncodeUnicode(str) {
-  // Firstly, escape the string using encodeURIComponent to get the UTF-8 encoding of the characters, 
+  // Firstly, escape the string using encodeURIComponent to get the UTF-8 encoding of the characters,
   // Secondly, we convert the percent encodings into raw bytes, and add it to btoa() function.
   const utf8Bytes = encodeURIComponent(str).replace(/%([0-9A-F]{2})/g, function (match, p1) {
     return String.fromCharCode('0x' + p1);
@@ -536,7 +497,7 @@ function base64EncodeUnicode(str) {
 async function notify(message) {
   const options = await this.getOptions();
   // message for initial clipping of the dom
-  if (message.type == "clip") {
+  if (message.type == 'clip') {
     // get the article info from the passed in dom
     const article = await getArticleFromDom(message.dom);
 
@@ -545,7 +506,7 @@ async function notify(message) {
     if (message.selection && message.clipSelection) {
       article.content = message.selection;
     }
-    
+
     // convert the article to markdown
     const { markdown, imageList } = await convertArticleToMarkdown(article);
 
@@ -556,40 +517,34 @@ async function notify(message) {
     const mdClipsFolder = await formatMdClipsFolder(article);
 
     // display the data in the popup
-    await browser.runtime.sendMessage({ type: "display.md", markdown: markdown, article: article, imageList: imageList, mdClipsFolder: mdClipsFolder});
+    await browser.runtime.sendMessage({ type: 'display.md', markdown: markdown, article: article, imageList: imageList, mdClipsFolder: mdClipsFolder });
   }
   // message for triggering download
-  else if (message.type == "download") {
+  else if (message.type == 'download') {
     downloadMarkdown(message.markdown, message.title, message.tab.id, message.imageList, message.mdClipsFolder);
   }
 }
 
 browser.commands.onCommand.addListener(function (command) {
-  const tab = browser.tabs.getCurrent()
-  if (command == "download_tab_as_markdown") {
-    const info = { menuItemId: "download-markdown-all" };
+  const tab = browser.tabs.getCurrent();
+  if (command == 'download_tab_as_markdown') {
+    const info = { menuItemId: 'download-markdown-all' };
     downloadMarkdownFromContext(info, tab);
-  }
-  else if (command == "copy_tab_as_markdown") {
-    const info = { menuItemId: "copy-markdown-all" };
+  } else if (command == 'copy_tab_as_markdown') {
+    const info = { menuItemId: 'copy-markdown-all' };
     copyMarkdownFromContext(info, tab);
-  }
-  else if (command == "copy_selection_as_markdown") {
-    const info = { menuItemId: "copy-markdown-selection" };
+  } else if (command == 'copy_selection_as_markdown') {
+    const info = { menuItemId: 'copy-markdown-selection' };
     copyMarkdownFromContext(info, tab);
-  }
-  else if (command == "copy_tab_as_markdown_link") {
+  } else if (command == 'copy_tab_as_markdown_link') {
     copyTabAsMarkdownLink(tab);
-  }
-  else if (command == "copy_selected_tab_as_markdown_link") {
+  } else if (command == 'copy_selected_tab_as_markdown_link') {
     copySelectedTabAsMarkdownLink(tab);
-  }
-  else if (command == "copy_selection_to_obsidian") {
-    const info = { menuItemId: "copy-markdown-obsidian" };
+  } else if (command == 'copy_selection_to_obsidian') {
+    const info = { menuItemId: 'copy-markdown-obsidian' };
     copyMarkdownFromContext(info, tab);
-  }
-  else if (command == "copy_tab_to_obsidian") {
-    const info = { menuItemId: "copy-markdown-obsall" };
+  } else if (command == 'copy_tab_to_obsidian') {
+    const info = { menuItemId: 'copy-markdown-obsall' };
     copyMarkdownFromContext(info, tab);
   }
 });
@@ -597,29 +552,27 @@ browser.commands.onCommand.addListener(function (command) {
 // click handler for the context menus
 browser.contextMenus.onClicked.addListener(function (info, tab) {
   // one of the copy to clipboard commands
-  if (info.menuItemId.startsWith("copy-markdown")) {
+  if (info.menuItemId.startsWith('copy-markdown')) {
     copyMarkdownFromContext(info, tab);
-  }
-  else if (info.menuItemId == "download-markdown-alltabs" || info.menuItemId == "tab-download-markdown-alltabs") {
+  } else if (info.menuItemId == 'download-markdown-alltabs' || info.menuItemId == 'tab-download-markdown-alltabs') {
     downloadMarkdownForAllTabs(info);
   }
   // one of the download commands
-  else if (info.menuItemId.startsWith("download-markdown")) {
+  else if (info.menuItemId.startsWith('download-markdown')) {
     downloadMarkdownFromContext(info, tab);
   }
   // copy tab as markdown link
-  else if (info.menuItemId.startsWith("copy-tab-as-markdown-link-all")) {
+  else if (info.menuItemId.startsWith('copy-tab-as-markdown-link-all')) {
     copyTabAsMarkdownLinkAll(tab);
   }
   // copy only selected tab as markdown link
-  else if (info.menuItemId.startsWith("copy-tab-as-markdown-link-selected")) {
+  else if (info.menuItemId.startsWith('copy-tab-as-markdown-link-selected')) {
     copySelectedTabAsMarkdownLink(tab);
-  }
-  else if (info.menuItemId.startsWith("copy-tab-as-markdown-link")) {
+  } else if (info.menuItemId.startsWith('copy-tab-as-markdown-link')) {
     copyTabAsMarkdownLink(tab);
   }
   // a settings toggle command
-  else if (info.menuItemId.startsWith("toggle-") || info.menuItemId.startsWith("tabtoggle-")) {
+  else if (info.menuItemId.startsWith('toggle-') || info.menuItemId.startsWith('tabtoggle-')) {
     toggleSetting(info.menuItemId.split('-')[1]);
   }
 });
@@ -628,45 +581,44 @@ browser.contextMenus.onClicked.addListener(function (info, tab) {
 async function toggleSetting(setting, options = null) {
   // if there's no options object passed in, we need to go get one
   if (options == null) {
-      // get the options from storage and toggle the setting
-      await toggleSetting(setting, await getOptions());
-  }
-  else {
+    // get the options from storage and toggle the setting
+    await toggleSetting(setting, await getOptions());
+  } else {
     // toggle the option and save back to storage
     options[setting] = !options[setting];
     await browser.storage.sync.set(options);
-    if (setting == "includeTemplate") {
-      browser.contextMenus.update("toggle-includeTemplate", {
-        checked: options.includeTemplate
+    if (setting == 'includeTemplate') {
+      browser.contextMenus.update('toggle-includeTemplate', {
+        checked: options.includeTemplate,
       });
       try {
-        browser.contextMenus.update("tabtoggle-includeTemplate", {
-          checked: options.includeTemplate
+        browser.contextMenus.update('tabtoggle-includeTemplate', {
+          checked: options.includeTemplate,
         });
-      } catch { }
+      } catch {}
     }
-    
-    if (setting == "downloadImages") {
-      browser.contextMenus.update("toggle-downloadImages", {
-        checked: options.downloadImages
+
+    if (setting == 'downloadImages') {
+      browser.contextMenus.update('toggle-downloadImages', {
+        checked: options.downloadImages,
       });
       try {
-        browser.contextMenus.update("tabtoggle-downloadImages", {
-          checked: options.downloadImages
+        browser.contextMenus.update('tabtoggle-downloadImages', {
+          checked: options.downloadImages,
         });
-      } catch { }
+      } catch {}
     }
   }
 }
 
 // this function ensures the content script is loaded (and loads it if it isn't)
 async function ensureScripts(tabId) {
-  const results = await browser.tabs.executeScript(tabId, { code: "typeof getSelectionAndDom === 'function';" })
+  const results = await browser.tabs.executeScript(tabId, { code: "typeof getSelectionAndDom === 'function';" });
   // The content script's last expression will be true if the function
   // has been defined. If this is not the case, then we need to run
   // pageScraper.js to define function getSelectionAndDom.
   if (!results || results[0] !== true) {
-    await browser.tabs.executeScript(tabId, {file: "/contentScript/contentScript.js"});
+    await browser.tabs.executeScript(tabId, { file: '/contentScript/contentScript.js' });
   }
 }
 
@@ -674,10 +626,10 @@ async function ensureScripts(tabId) {
 async function getArticleFromDom(domString) {
   // parse the dom
   const parser = new DOMParser();
-  const dom = parser.parseFromString(domString, "text/html");
+  const dom = parser.parseFromString(domString, 'text/html');
 
-  if (dom.documentElement.nodeName == "parsererror") {
-    console.error("error while parsing");
+  if (dom.documentElement.nodeName == 'parsererror') {
+    console.error('error while parsing');
   }
 
   const math = {};
@@ -689,72 +641,72 @@ async function getArticleFromDom(domString) {
     math[randomId] = mathInfo;
   };
 
-  dom.body.querySelectorAll('script[id^=MathJax-Element-]')?.forEach(mathSource => {
-    const type = mathSource.attributes.type.value
+  dom.body.querySelectorAll('script[id^=MathJax-Element-]')?.forEach((mathSource) => {
+    const type = mathSource.attributes.type.value;
     storeMathInfo(mathSource, {
       tex: mathSource.innerText,
-      inline: type ? !type.includes('mode=display') : false
+      inline: type ? !type.includes('mode=display') : false,
     });
   });
 
-  dom.body.querySelectorAll('[markdownload-latex]')?.forEach(mathJax3Node =>  {
-    const tex = mathJax3Node.getAttribute('markdownload-latex')
-    const display = mathJax3Node.getAttribute('display')
-    const inline = !(display && display === 'true')
+  dom.body.querySelectorAll('[markdownload-latex]')?.forEach((mathJax3Node) => {
+    const tex = mathJax3Node.getAttribute('markdownload-latex');
+    const display = mathJax3Node.getAttribute('display');
+    const inline = !(display && display === 'true');
 
-    const mathNode = document.createElement(inline ? "i" : "p")
+    const mathNode = document.createElement(inline ? 'i' : 'p');
     mathNode.textContent = tex;
-    mathJax3Node.parentNode.insertBefore(mathNode, mathJax3Node.nextSibling)
-    mathJax3Node.parentNode.removeChild(mathJax3Node)
+    mathJax3Node.parentNode.insertBefore(mathNode, mathJax3Node.nextSibling);
+    mathJax3Node.parentNode.removeChild(mathJax3Node);
 
     storeMathInfo(mathNode, {
       tex: tex,
-      inline: inline
+      inline: inline,
     });
   });
 
-  dom.body.querySelectorAll('.katex-mathml')?.forEach(kaTeXNode => {
+  dom.body.querySelectorAll('.katex-mathml')?.forEach((kaTeXNode) => {
     storeMathInfo(kaTeXNode, {
       tex: kaTeXNode.querySelector('annotation').textContent,
-      inline: true
+      inline: true,
     });
   });
 
-  dom.body.querySelectorAll('[class*=highlight-text],[class*=highlight-source]')?.forEach(codeSource => {
-    const language = codeSource.className.match(/highlight-(?:text|source)-([a-z0-9]+)/)?.[1]
-    if (codeSource.firstChild.nodeName == "PRE") {
-      codeSource.firstChild.id = `code-lang-${language}`
+  dom.body.querySelectorAll('[class*=highlight-text],[class*=highlight-source]')?.forEach((codeSource) => {
+    const language = codeSource.className.match(/highlight-(?:text|source)-([a-z0-9]+)/)?.[1];
+    if (codeSource.firstChild.nodeName == 'PRE') {
+      codeSource.firstChild.id = `code-lang-${language}`;
     }
   });
 
-  dom.body.querySelectorAll('[class*=language-]')?.forEach(codeSource => {
-    const language = codeSource.className.match(/language-([a-z0-9]+)/)?.[1]
+  dom.body.querySelectorAll('[class*=language-]')?.forEach((codeSource) => {
+    const language = codeSource.className.match(/language-([a-z0-9]+)/)?.[1];
     codeSource.id = `code-lang-${language}`;
   });
 
-  dom.body.querySelectorAll('pre br')?.forEach(br => {
+  dom.body.querySelectorAll('pre br')?.forEach((br) => {
     // we need to keep <br> tags because they are removed by Readability.js
     br.outerHTML = '<br-keep></br-keep>';
   });
 
-  dom.body.querySelectorAll('.codehilite > pre')?.forEach(codeSource => {
+  dom.body.querySelectorAll('.codehilite > pre')?.forEach((codeSource) => {
     if (codeSource.firstChild.nodeName !== 'CODE' && !codeSource.className.includes('language')) {
       codeSource.id = `code-lang-text`;
     }
   });
 
-  dom.body.querySelectorAll('h1, h2, h3, h4, h5, h6')?.forEach(header => {
+  dom.body.querySelectorAll('h1, h2, h3, h4, h5, h6')?.forEach((header) => {
     // Readability.js will strip out headings from the dom if certain words appear in their className
-    // See: https://github.com/mozilla/readability/issues/807  
+    // See: https://github.com/mozilla/readability/issues/807
     header.className = '';
-    header.outerHTML = header.outerHTML;  
+    header.outerHTML = header.outerHTML;
   });
 
   // Prevent Readability from removing the <html> element if has a 'class' attribute
   // which matches removal criteria.
   // Note: The document element is guaranteed to be the HTML tag because the 'text/html'
   // mime type was used when the DOM was created.
-  dom.documentElement.removeAttribute('class')
+  dom.documentElement.removeAttribute('class');
 
   // simplify the dom into an article
   const article = new Readability(dom).parse();
@@ -773,24 +725,26 @@ async function getArticleFromDom(domString) {
   article.port = url.port;
   article.protocol = url.protocol;
   article.search = url.search;
-  
 
   // make sure the dom has a head
   if (dom.head) {
     // and the keywords, should they exist, as an array
-    article.keywords = dom.head.querySelector('meta[name="keywords"]')?.content?.split(',')?.map(s => s.trim());
+    article.keywords = dom.head
+      .querySelector('meta[name="keywords"]')
+      ?.content?.split(',')
+      ?.map((s) => s.trim());
 
     // add all meta tags, so users can do whatever they want
-    dom.head.querySelectorAll('meta[name][content], meta[property][content]')?.forEach(meta => {
-      const key = (meta.getAttribute('name') || meta.getAttribute('property'))
-      const val = meta.getAttribute('content')
+    dom.head.querySelectorAll('meta[name][content], meta[property][content]')?.forEach((meta) => {
+      const key = meta.getAttribute('name') || meta.getAttribute('property');
+      const val = meta.getAttribute('content');
       if (key && val && !article[key]) {
         article[key] = val;
       }
-    })
+    });
   }
 
-  article.math = math
+  article.math = math;
 
   // return the article
   return article;
@@ -800,7 +754,7 @@ async function getArticleFromDom(domString) {
 // `selection` is a bool indicating whether we should just get the selected text
 async function getArticleFromContent(tabId, selection = false) {
   // run the content script function to get the details
-  const results = await browser.tabs.executeScript(tabId, { code: "getSelectionAndDom()" });
+  const results = await browser.tabs.executeScript(tabId, { code: 'getSelectionAndDom()' });
 
   // make sure we actually got a valid result
   if (results && results[0] && results[0].dom) {
@@ -814,16 +768,18 @@ async function getArticleFromContent(tabId, selection = false) {
 
     //return the article
     return article;
-  }
-  else return null;
+  } else return null;
 }
 
 // function to apply the title template
 async function formatTitle(article) {
   let options = await getOptions();
-  
+
   let title = textReplace(options.title, article, options.disallowedChars + '/');
-  title = title.split('/').map(s=>generateValidFileName(s, options.disallowedChars)).join('/');
+  title = title
+    .split('/')
+    .map((s) => generateValidFileName(s, options.disallowedChars))
+    .join('/');
   return title;
 }
 
@@ -833,7 +789,10 @@ async function formatMdClipsFolder(article) {
   let mdClipsFolder = '';
   if (options.mdClipsFolder && options.downloadMode == 'downloadsApi') {
     mdClipsFolder = textReplace(options.mdClipsFolder, article, options.disallowedChars);
-    mdClipsFolder = mdClipsFolder.split('/').map(s => generateValidFileName(s, options.disallowedChars)).join('/');
+    mdClipsFolder = mdClipsFolder
+      .split('/')
+      .map((s) => generateValidFileName(s, options.disallowedChars))
+      .join('/');
     if (!mdClipsFolder.endsWith('/')) mdClipsFolder += '/';
   }
 
@@ -846,7 +805,10 @@ async function formatObsidianFolder(article) {
   let obsidianFolder = '';
   if (options.obsidianFolder) {
     obsidianFolder = textReplace(options.obsidianFolder, article, options.disallowedChars);
-    obsidianFolder = obsidianFolder.split('/').map(s => generateValidFileName(s, options.disallowedChars)).join('/');
+    obsidianFolder = obsidianFolder
+      .split('/')
+      .map((s) => generateValidFileName(s, options.disallowedChars))
+      .join('/');
     if (!obsidianFolder.endsWith('/')) obsidianFolder += '/';
   }
 
@@ -856,13 +818,12 @@ async function formatObsidianFolder(article) {
 // function to download markdown, triggered by context menu
 async function downloadMarkdownFromContext(info, tab) {
   await ensureScripts(tab.id);
-  const article = await getArticleFromContent(tab.id, info.menuItemId == "download-markdown-selection");
+  const article = await getArticleFromContent(tab.id, info.menuItemId == 'download-markdown-selection');
   const title = await formatTitle(article);
   const { markdown, imageList } = await convertArticleToMarkdown(article);
   // format the mdClipsFolder
   const mdClipsFolder = await formatMdClipsFolder(article);
-  await downloadMarkdown(markdown, title, tab.id, imageList, mdClipsFolder); 
-
+  await downloadMarkdown(markdown, title, tab.id, imageList, mdClipsFolder);
 }
 
 // function to copy a tab url as a markdown link
@@ -873,12 +834,11 @@ async function copyTabAsMarkdownLink(tab) {
     const title = await formatTitle(article);
     await browser.tabs.executeScript(tab.id, { code: `copyToClipboard("[${title}](${article.baseURI})")` });
     // await navigator.clipboard.writeText(`[${title}](${article.baseURI})`);
-  }
-  catch (error) {
+  } catch (error) {
     // This could happen if the extension is not allowed to run code in
     // the page, for example if the tab is a privileged page.
-    console.error("Failed to copy as markdown link: " + error);
-  };
+    console.error('Failed to copy as markdown link: ' + error);
+  }
 }
 
 // function to copy all tabs as markdown links
@@ -887,27 +847,25 @@ async function copyTabAsMarkdownLinkAll(tab) {
     const options = await getOptions();
     options.frontmatter = options.backmatter = '';
     const tabs = await browser.tabs.query({
-      currentWindow: true
+      currentWindow: true,
     });
-    
+
     const links = [];
-    for(const tab of tabs) {
+    for (const tab of tabs) {
       await ensureScripts(tab.id);
       const article = await getArticleFromContent(tab.id);
       const title = await formatTitle(article);
-      const link = `${options.bulletListMarker} [${title}](${article.baseURI})`
-      links.push(link)
-    };
-    
-    const markdown = links.join(`\n`)
-    await browser.tabs.executeScript(tab.id, { code: `copyToClipboard(${JSON.stringify(markdown)})` });
+      const link = `${options.bulletListMarker} [${title}](${article.baseURI})`;
+      links.push(link);
+    }
 
-  }
-  catch (error) {
+    const markdown = links.join(`\n`);
+    await browser.tabs.executeScript(tab.id, { code: `copyToClipboard(${JSON.stringify(markdown)})` });
+  } catch (error) {
     // This could happen if the extension is not allowed to run code in
     // the page, for example if the tab is a privileged page.
-    console.error("Failed to copy as markdown link: " + error);
-  };
+    console.error('Failed to copy as markdown link: ' + error);
+  }
 }
 
 // function to copy only selected tabs as markdown links
@@ -917,7 +875,7 @@ async function copySelectedTabAsMarkdownLink(tab) {
     options.frontmatter = options.backmatter = '';
     const tabs = await browser.tabs.query({
       currentWindow: true,
-      highlighted: true
+      highlighted: true,
     });
 
     const links = [];
@@ -925,82 +883,75 @@ async function copySelectedTabAsMarkdownLink(tab) {
       await ensureScripts(tab.id);
       const article = await getArticleFromContent(tab.id);
       const title = await formatTitle(article);
-      const link = `${options.bulletListMarker} [${title}](${article.baseURI})`
-      links.push(link)
-    };
+      const link = `${options.bulletListMarker} [${title}](${article.baseURI})`;
+      links.push(link);
+    }
 
-    const markdown = links.join(`\n`)
+    const markdown = links.join(`\n`);
     await browser.tabs.executeScript(tab.id, { code: `copyToClipboard(${JSON.stringify(markdown)})` });
-
-  }
-  catch (error) {
+  } catch (error) {
     // This could happen if the extension is not allowed to run code in
     // the page, for example if the tab is a privileged page.
-    console.error("Failed to copy as markdown link: " + error);
-  };
+    console.error('Failed to copy as markdown link: ' + error);
+  }
 }
 
 // function to copy markdown to the clipboard, triggered by context menu
 async function copyMarkdownFromContext(info, tab) {
-  try{
+  try {
     await ensureScripts(tab.id);
 
     const platformOS = navigator.platform;
-    var folderSeparator = "";
-    if(platformOS.indexOf("Win") === 0){
-      folderSeparator = "\\";
-    }else{
-      folderSeparator = "/";
+    var folderSeparator = '';
+    if (platformOS.indexOf('Win') === 0) {
+      folderSeparator = '\\';
+    } else {
+      folderSeparator = '/';
     }
 
-    if (info.menuItemId == "copy-markdown-link") {
+    if (info.menuItemId == 'copy-markdown-link') {
       const options = await getOptions();
       options.frontmatter = options.backmatter = '';
       const article = await getArticleFromContent(tab.id, false);
       const { markdown } = turndown(`<a href="${info.linkUrl}">${info.linkText || info.selectionText}</a>`, { ...options, downloadImages: false }, article);
-      await browser.tabs.executeScript(tab.id, {code: `copyToClipboard(${JSON.stringify(markdown)})`});
-    }
-    else if (info.menuItemId == "copy-markdown-image") {
-      await browser.tabs.executeScript(tab.id, {code: `copyToClipboard("![](${info.srcUrl})")`});
-    }
-    else if(info.menuItemId == "copy-markdown-obsidian") {
-      const article = await getArticleFromContent(tab.id, info.menuItemId == "copy-markdown-obsidian");
+      await browser.tabs.executeScript(tab.id, { code: `copyToClipboard(${JSON.stringify(markdown)})` });
+    } else if (info.menuItemId == 'copy-markdown-image') {
+      await browser.tabs.executeScript(tab.id, { code: `copyToClipboard("![](${info.srcUrl})")` });
+    } else if (info.menuItemId == 'copy-markdown-obsidian') {
+      const article = await getArticleFromContent(tab.id, info.menuItemId == 'copy-markdown-obsidian');
       const title = await formatTitle(article);
       const options = await getOptions();
       const obsidianVault = options.obsidianVault;
       const obsidianFolder = await formatObsidianFolder(article);
-      const { markdown } = await convertArticleToMarkdown(article, downloadImages = false);
+      const { markdown } = await convertArticleToMarkdown(article, (downloadImages = false));
       await browser.tabs.executeScript(tab.id, { code: `copyToClipboard(${JSON.stringify(markdown)})` });
-      await chrome.tabs.update({url: "obsidian://advanced-uri?vault=" + obsidianVault + "&clipboard=true&mode=new&filepath=" + obsidianFolder + generateValidFileName(title)});
-    }
-    else if(info.menuItemId == "copy-markdown-obsall") {
-      const article = await getArticleFromContent(tab.id, info.menuItemId == "copy-markdown-obsall");
+      await chrome.tabs.update({ url: 'obsidian://advanced-uri?vault=' + obsidianVault + '&clipboard=true&mode=new&filepath=' + obsidianFolder + generateValidFileName(title) });
+    } else if (info.menuItemId == 'copy-markdown-obsall') {
+      const article = await getArticleFromContent(tab.id, info.menuItemId == 'copy-markdown-obsall');
       const title = await formatTitle(article);
       const options = await getOptions();
       const obsidianVault = options.obsidianVault;
       const obsidianFolder = await formatObsidianFolder(article);
-      const { markdown } = await convertArticleToMarkdown(article, downloadImages = false);
+      const { markdown } = await convertArticleToMarkdown(article, (downloadImages = false));
       await browser.tabs.executeScript(tab.id, { code: `copyToClipboard(${JSON.stringify(markdown)})` });
-      await browser.tabs.update({url: "obsidian://advanced-uri?vault=" + obsidianVault + "&clipboard=true&mode=new&filepath=" + obsidianFolder + generateValidFileName(title)});
-    }
-    else {
-      const article = await getArticleFromContent(tab.id, info.menuItemId == "copy-markdown-selection");
-      const { markdown } = await convertArticleToMarkdown(article, downloadImages = false);
+      await browser.tabs.update({ url: 'obsidian://advanced-uri?vault=' + obsidianVault + '&clipboard=true&mode=new&filepath=' + obsidianFolder + generateValidFileName(title) });
+    } else {
+      const article = await getArticleFromContent(tab.id, info.menuItemId == 'copy-markdown-selection');
+      const { markdown } = await convertArticleToMarkdown(article, (downloadImages = false));
       await browser.tabs.executeScript(tab.id, { code: `copyToClipboard(${JSON.stringify(markdown)})` });
     }
-  }
-  catch (error) {
+  } catch (error) {
     // This could happen if the extension is not allowed to run code in
     // the page, for example if the tab is a privileged page.
-    console.error("Failed to copy text: " + error);
-  };
+    console.error('Failed to copy text: ' + error);
+  }
 }
 
 async function downloadMarkdownForAllTabs(info) {
   const tabs = await browser.tabs.query({
-    currentWindow: true
+    currentWindow: true,
   });
-  tabs.forEach(tab => {
+  tabs.forEach((tab) => {
     downloadMarkdownFromContext(info, tab);
   });
 }
@@ -1012,15 +963,13 @@ async function downloadMarkdownForAllTabs(info) {
  * @license MIT
  */
 if (!String.prototype.replaceAll) {
-	String.prototype.replaceAll = function(str, newStr){
+  String.prototype.replaceAll = function (str, newStr) {
+    // If a regex pattern
+    if (Object.prototype.toString.call(str).toLowerCase() === '[object regexp]') {
+      return this.replace(str, newStr);
+    }
 
-		// If a regex pattern
-		if (Object.prototype.toString.call(str).toLowerCase() === '[object regexp]') {
-			return this.replace(str, newStr);
-		}
-
-		// If a string
-		return this.replace(new RegExp(str, 'g'), newStr);
-
-	};
+    // If a string
+    return this.replace(new RegExp(str, 'g'), newStr);
+  };
 }